refactoring basic common code path still too many v6 not handling slash
