+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pnpm/lockfile/PnpmLockExtractor.java
@@ -9,17 +9,17 @@ import org.jetbrains.annotations.Nullable;
 
 import com.synopsys.integration.detectable.detectable.codelocation.CodeLocation;
 import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLinkedPackageResolver;
-import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLockYamlParser;
+import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLockYamlParserInitial;
 import com.synopsys.integration.detectable.detectables.yarn.packagejson.NullSafePackageJson;
 import com.synopsys.integration.detectable.detectables.yarn.packagejson.PackageJsonFiles;
 import com.synopsys.integration.detectable.extraction.Extraction;
 import com.synopsys.integration.util.NameVersion;
 
 public class PnpmLockExtractor {
-    private final PnpmLockYamlParser pnpmLockYamlParser;
+    private final PnpmLockYamlParserInitial pnpmLockYamlParser;
     private final PackageJsonFiles packageJsonFiles;
 
-    public PnpmLockExtractor(PnpmLockYamlParser pnpmLockYamlParser, PackageJsonFiles packageJsonFiles) {
+    public PnpmLockExtractor(PnpmLockYamlParserInitial pnpmLockYamlParser, PackageJsonFiles packageJsonFiles) {
         this.pnpmLockYamlParser = pnpmLockYamlParser;
         this.packageJsonFiles = packageJsonFiles;
     }
