+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pnpm/lockfile/model/PnpmLockYamlBase.java
@@ -1,0 +1,8 @@
+package com.synopsys.integration.detectable.detectables.pnpm.lockfile.model;
+
+import org.jetbrains.annotations.Nullable;
+
+public class PnpmLockYamlBase {
+    @Nullable
+    public String lockfileVersion;
+}
