+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pnpm/lockfile/process/PnpmLockYamlParser.java
@@ -1,76 +1,94 @@
 package com.synopsys.integration.detectable.detectables.pnpm.lockfile.process;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
+import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 
+import org.apache.commons.collections4.MapUtils;
+import org.apache.commons.collections4.Predicate;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
-import org.yaml.snakeyaml.DumperOptions;
-import org.yaml.snakeyaml.LoaderOptions;
-import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.constructor.Constructor;
-import org.yaml.snakeyaml.constructor.ConstructorException;
-import org.yaml.snakeyaml.representer.Representer;
 
 import com.synopsys.integration.detectable.detectable.codelocation.CodeLocation;
-import com.synopsys.integration.detectable.detectable.util.EnumListFilter;
-import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmDependencyType;
 import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmLockYaml;
-import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmLockYamlv5;
-import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmLockYamlv6;
+import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmProjectPackagev6;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.util.NameVersion;
 
 public class PnpmLockYamlParser {
+    private static final Predicate<String> isNodeRoot = "."::equals;
 
-    private final EnumListFilter<PnpmDependencyType> dependencyFilter;
+    private PnpmYamlTransformerv6 pnpmTransformer;
 
-    public PnpmLockYamlParser(EnumListFilter<PnpmDependencyType> dependencyFilter) {
-        this.dependencyFilter = dependencyFilter;
+    public PnpmLockYamlParser(PnpmYamlTransformerv6 pnpmTransformer) {
+        this.pnpmTransformer = pnpmTransformer;
     }
 
-    public List<CodeLocation> parse(File pnpmLockYamlFile, @Nullable NameVersion projectNameVersion, PnpmLinkedPackageResolver linkedPackageResolver)
-        throws IOException, IntegrationException {
-        PnpmLockYaml pnpmLockYaml = parseYamlFile(pnpmLockYamlFile);
-        
-        if (pnpmLockYaml instanceof PnpmLockYamlv6) {
-            PnpmYamlTransformerv6 pnpmYamlTransformer = new PnpmYamlTransformerv6(dependencyFilter);
-            PnpmLockYamlParserv6 pnpmYamlParser = new PnpmLockYamlParserv6(pnpmYamlTransformer);
-            return pnpmYamlParser.parse(pnpmLockYamlFile.getParentFile(), (PnpmLockYamlv6) pnpmLockYaml, linkedPackageResolver, projectNameVersion);
-        } else {
-            PnpmYamlTransformerv5 pnpmYamlTransformer = new PnpmYamlTransformerv5(dependencyFilter);
-            PnpmLockYamlParserv5 pnpmYamlParser = new PnpmLockYamlParserv5(pnpmYamlTransformer);
-            return pnpmYamlParser.parse(pnpmLockYamlFile.getParentFile(), (PnpmLockYamlv5) pnpmLockYaml, linkedPackageResolver, projectNameVersion);
+    public List<CodeLocation> parse(File parentFile, PnpmLockYaml pnpmLockYaml,
+            PnpmLinkedPackageResolver linkedPackageResolver, @Nullable NameVersion projectNameVersion)
+            throws IntegrationException {
+        List<CodeLocation> codeLocationsFromImports = createCodeLocationsFromImports(parentFile, pnpmLockYaml,
+                linkedPackageResolver, projectNameVersion);
+        if (codeLocationsFromImports.isEmpty()) {
+            return createCodeLocationsFromRoot(parentFile, pnpmLockYaml, projectNameVersion, linkedPackageResolver);
         }
+        return codeLocationsFromImports;
     }
 
-    /**
-     * This method reads the pnpm-lock.yaml. It first tries to read it in v6 format
-     * and then tries v5 if that fails. This is usually faster than first cracking
-     * open the yaml file, checking what version it is, and then calling the
-     * appropriate reader.
-     * 
-     * @param pnpmLockYamlFile the File path to the pnpm-lock.yaml file
-     * @return a memory representation of the lock file.
-     * @throws FileNotFoundException
-     */
-    private PnpmLockYaml parseYamlFile(File pnpmLockYamlFile) throws FileNotFoundException {
-        DumperOptions dumperOptions = new DumperOptions();
-        Representer representer = new Representer(dumperOptions);
-        representer.getPropertyUtils().setSkipMissingProperties(true);
+    private List<CodeLocation> createCodeLocationsFromRoot(File sourcePath, PnpmLockYaml pnpmLockYaml,
+            @Nullable NameVersion projectNameVersion, PnpmLinkedPackageResolver linkedPackageResolver)
+            throws IntegrationException {
+        CodeLocation codeLocation = pnpmTransformer.generateCodeLocation(sourcePath, pnpmLockYaml, projectNameVersion,
+                linkedPackageResolver);
+        return Collections.singletonList(codeLocation);
+    }
 
-        LoaderOptions loaderOptions = new LoaderOptions();
-        try {
-            // Try to read the lockfile into the v6 Yaml classes. It's more common and 
-            // should hopefully work more of the time.
-            Yaml yaml = new Yaml(new Constructor(PnpmLockYamlv6.class, loaderOptions), representer);
-            return yaml.load(new FileReader(pnpmLockYamlFile));
-        } catch (ConstructorException e) {
-            // If the v6 reading fails try to read a v5 Yaml. 
-            Yaml yaml = new Yaml(new Constructor(PnpmLockYamlv5.class, loaderOptions), representer);
-            return yaml.load(new FileReader(pnpmLockYamlFile));
+    private List<CodeLocation> createCodeLocationsFromImports(File sourcePath, PnpmLockYaml pnpmLockYaml,
+            PnpmLinkedPackageResolver linkedPackageResolver, @Nullable NameVersion projectNameVersion)
+            throws IntegrationException {
+        if (MapUtils.isEmpty(pnpmLockYaml.importers)) {
+            return Collections.emptyList();
         }
+
+        List<CodeLocation> codeLocations = new LinkedList<>();
+        for (Map.Entry<String, PnpmProjectPackagev6> projectPackageInfo : pnpmLockYaml.importers.entrySet()) {
+            String projectKey = projectPackageInfo.getKey();
+            PnpmProjectPackagev6 projectPackage = projectPackageInfo.getValue();
+            NameVersion extractedNameVersion = extractProjectInfo(projectPackageInfo, linkedPackageResolver,
+                    projectNameVersion);
+
+            String reportingProjectPackagePath = null;
+            if (!isNodeRoot.evaluate(projectKey)) {
+                reportingProjectPackagePath = projectKey;
+            }
+            File generatedSourcePath = generateCodeLocationSourcePath(sourcePath, reportingProjectPackagePath);
+
+            codeLocations.add(pnpmTransformer.generateCodeLocation(generatedSourcePath, projectPackage,
+                    reportingProjectPackagePath, extractedNameVersion, pnpmLockYaml.packages, linkedPackageResolver, pnpmLockYaml.snapshots));
+        }
+
+        return codeLocations;
+    }
+
+    private NameVersion extractProjectInfo(Map.Entry<String, PnpmProjectPackagev6> projectPackageInfo,
+            PnpmLinkedPackageResolver linkedPackageResolver, @Nullable NameVersion projectNameVersion) {
+        if (isNodeRoot.evaluate(projectPackageInfo.getKey()) && projectNameVersion != null
+                && projectNameVersion.getName() != null) {
+            // resolve "." package to project root
+            return projectNameVersion;
+        }
+
+        String projectPackageName = projectPackageInfo.getKey();
+        String projectPackageVersion = linkedPackageResolver.resolveVersionOfLinkedPackage(null, projectPackageName);
+        return new NameVersion(projectPackageName, projectPackageVersion);
+    }
+
+    private File generateCodeLocationSourcePath(File sourcePath, @Nullable String reportingProjectPackagePath) {
+        if (StringUtils.isNotEmpty(reportingProjectPackagePath)) {
+            return new File(sourcePath, reportingProjectPackagePath);
+        }
+        return sourcePath;
     }
 }
