+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/pnpm/functional/PnpmLockExtractorTest.java
@@ -10,7 +10,7 @@ import com.synopsys.integration.detectable.detectable.util.EnumListFilter;
 import com.synopsys.integration.detectable.detectables.pnpm.lockfile.PnpmLockExtractor;
 import com.synopsys.integration.detectable.detectables.pnpm.lockfile.model.PnpmDependencyType;
 import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLinkedPackageResolver;
-import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLockYamlParser;
+import com.synopsys.integration.detectable.detectables.pnpm.lockfile.process.PnpmLockYamlParserInitial;
 import com.synopsys.integration.detectable.detectables.yarn.packagejson.PackageJsonFiles;
 import com.synopsys.integration.detectable.detectables.yarn.packagejson.PackageJsonReader;
 import com.synopsys.integration.detectable.extraction.Extraction;
@@ -21,7 +21,7 @@ public class PnpmLockExtractorTest {
     public void testNoFailureOnNullPackageJsonv5() {
         PackageJsonFiles packageJsonFiles = new PackageJsonFiles(new PackageJsonReader(new Gson()));
         EnumListFilter<PnpmDependencyType> dependencyTypeFilter = EnumListFilter.excludeNone();
-        PnpmLockExtractor extractor = new PnpmLockExtractor(new PnpmLockYamlParser(dependencyTypeFilter), packageJsonFiles);
+        PnpmLockExtractor extractor = new PnpmLockExtractor(new PnpmLockYamlParserInitial(dependencyTypeFilter), packageJsonFiles);
 
         File pnpmLockYaml = FunctionalTestFiles.asFile("/pnpm/v5/pnpm-lock.yaml");
         Extraction extraction = extractor.extract(pnpmLockYaml, null, new PnpmLinkedPackageResolver(new File(""), packageJsonFiles));
@@ -32,7 +32,7 @@ public class PnpmLockExtractorTest {
     public void testNoFailureOnNullPackageJsonv6() {
         PackageJsonFiles packageJsonFiles = new PackageJsonFiles(new PackageJsonReader(new Gson()));
         EnumListFilter<PnpmDependencyType> dependencyTypeFilter = EnumListFilter.excludeNone();
-        PnpmLockExtractor extractor = new PnpmLockExtractor(new PnpmLockYamlParser(dependencyTypeFilter), packageJsonFiles);
+        PnpmLockExtractor extractor = new PnpmLockExtractor(new PnpmLockYamlParserInitial(dependencyTypeFilter), packageJsonFiles);
 
         File pnpmLockYaml = FunctionalTestFiles.asFile("/pnpm/v6/pnpm-lock.yaml");
         Extraction extraction = extractor.extract(pnpmLockYaml, null, new PnpmLinkedPackageResolver(new File(""), packageJsonFiles));
